+++ R/zzz.R
@@ -1,2 +1,6 @@
 
-utils::globalVariables(c("%>%", "filter", "label", "translation", ".datamods_edit_update", ".datamods_edit_delete", ".datamods_id", "..var_edit"))
+utils::globalVariables(c(
+  "%>%", "filter", "label", "translation",
+  ".datamods_edit_update", ".datamods_edit_delete", ".datamods_id",
+  "..var_edit", "..vars_datamods_edit"
+))